Restructured model folders to highlight related parent and child classes
