+++ Antlers.UnitTests/Sleeper/Mocks/LeagueFaker.cs
@@ -1,4 +1,5 @@
 ﻿using Antlers.Sleeper.Models;
+using Antlers.Sleeper.Models.League;
 using Bogus;
 
 namespace Antlers.UnitTests.Sleeper.Mocks
