+++ Antlers.UnitTests/Sleeper/Mocks/RosterFaker.cs
@@ -1,4 +1,5 @@
 using Antlers.Sleeper.Models;
+using Antlers.Sleeper.Models.Roster;
 using Bogus;
 
 namespace Antlers.UnitTests.Sleeper.Mocks;
