+++ Antlers/Sleeper/Models/League/ScoringSettings.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace Antlers.Sleeper.Models
+namespace Antlers.Sleeper.Models.League
 {
     public class ScoringSettings
     {
